reafactor: signal trap
